pyproject.toml
[project]
name = "atss"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.74.1",
    "beautifulsoup4>=4.14.2",
    "dateparser>=1.2.2",
    "dotenv>=0.9.9",
    "feedparser>=6.0.12",
    "fuzzywuzzy>=0.18.0",
    "httpx[socks]>=0.28.1",
    "jinja2>=3.1.6",
    "listparser>=0.20",
    "lxml>=6.0.2",
    "markdown>=3.10",
    "nltk>=3.9.2",
    "numpy>=2.3.5",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "scrapy>=2.13.4",
    "selenium>=4.38.0",
    "spacy>=3.8.11",
    "sqlalchemy>=2.0.44",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
]

[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"
